Phase 6: rewrite .kpx to v8 layout and fix dedup u8 saturation
Rework the .kpx codec for higher search throughput and correctness.
All four index files share format_version 8 and the kix/kpx magics
become "KIX8" / "KPX8".

.kpx v8 layout
  - Per-kmer payload is split into a 5×u32 top header + a 2-bit kind
    map (indexed by the .kix decoded distinct seq_id array) + partition
    groups (no per-group seq_id) + occ=1 / occ>=2 split short buckets.
  - The seq_id is no longer stored in the .kpx posting; the decoder
    runs a 2-pointer merge walk between the .kix decoded array and the
    sorted candidate list, resolving each candidate's (kind, rank)
    pair via the kind map.
  - FOR-block header is widened to 8 B ([u32 min][u8 b][3 B pad])
    so the body starts on an 8 B boundary; stream tails switch from a
    varint stream to a packed bit-width form
    ([u8 count][u32 min][u8 b][bitpacked body]).
  - PosDecodeScratch (per-thread, reusable) holds the per-kind decoded
    position buffers so the search hot path avoids per-kmer allocation.

Search wiring
  - PosDecoder ctor takes (data, end, kix_decoded, kix_count,
    candidates, n_candidates, scratch).
  - Stage 2 re-decodes the .kix posting per k-mer with SeqIdDecoder
    and feeds decoded_data()/decoded_count() into PosDecoder.  A
    thread-local PosDecodeScratch is reused across calls.
  - Stage 2 fast-paths: skip k-mers whose kix posting is empty, and
    bail out early in PosDecoder when kix_count == 0 (avoids touching
    the .kpx posting at offset 0, which production builders alias for
    every empty k-mer).

dedup u8 saturation fix (correctness)
  - seq_id_dedup::dedup_seq_ids now returns u32 occ_count_out.  The
    prior u8 path silently saturated runs > 255, which made
    encode_posting_kpx advance pos_cursor by the saturated value and
    drop the trailing positions.  In nt_v4 vol 75 this lost about
    91.6 M positions (2.3% of the volume).  Mode 1 search did not see
    the corruption because it never opens the .kpx; mode 2/3 hits
    were silently misassigned.
  - encode_posting_kpx and KpxWriter now take u32 occ_count; the wire
    format is unchanged (partition occ_count was already u32 on disk;
    short_occ_ge2 occ_count remains u8 because freq_threshold_part is
    capped at 255 by the CLI).  Regression tests cover both the
    dedup kernel (test_run_length_above_255) and the round-trip
    through encode_posting_kpx (test_partition_above_255 with 1000
    occurrences in one cluster).

Tools
  - tools/index_compress_stats: rewrite the walker for the v8 layout
    and require --kix so empty k-mers (whose .kpx pos_offset is 0 and
    aliases the first non-empty payload in production indexes) are
    correctly skipped.
  - tools/dump_postings: pass the .kix decoded array into the new
    candidate-set decoder API.

Tests
  - test/test_kpx_io: rewrite for v8 with cases for all-occ=1, mixed
    short buckets, partition only, FOR tail sweep (0/1/127), 2-block
    short_occ1 stream, distinct_count==0 empty posting, all-candidate
    miss, and the > 255 occurrence regression.
  - test/test_seq_id_dedup: u8 -> u32 occ buffers and add the
    regression for 256 / 1000 / 100 000 long runs.
  - test/test_builder: switch the helpers to the v8 PosDecoder API
    and recompute kpx_position_count from the v8 top header.

CLAUDE.md and the kix/kpx/ksx/khx format headers are updated to
describe the v8 layout and the Phase 6 lineage.  IKAFSSN_VERSION
moves to 0.1.2026.05.01.

+++ src/index/index_builder.cpp
@@ -254,11 +254,11 @@ bool build_index(BlastDbReader& db,
     uint64_t reserve_entries = config.memory_limit / parallel_sort_entry_overhead();
 
     // Process each partition (sequentially to respect memory constraints)
-    // Buffers reused per k-mer for v7 dedup + encoder pipeline.
+    // Buffers reused per k-mer for the v8 dedup + encoder pipeline.
     std::vector<uint32_t> sid_buf;          sid_buf.reserve(64 * 1024);
     std::vector<uint32_t> abs_pos_buf;      abs_pos_buf.reserve(64 * 1024);
     std::vector<uint32_t> distinct_sid_buf; distinct_sid_buf.reserve(64 * 1024);
-    std::vector<uint8_t>  occ_count_buf;    occ_count_buf.reserve(64 * 1024);
+    std::vector<uint32_t> occ_count_buf;    occ_count_buf.reserve(64 * 1024);
     std::vector<uint32_t> seq_delta_buf;    seq_delta_buf.reserve(64 * 1024);
     std::vector<uint8_t>  pfd_out_buf;      pfd_out_buf.reserve(256 * 1024);
